pyproject.toml
[project]
name = "pipecat-modal"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "bs4>=0.0.2",
    "llama-index==0.12.41",
    "llama-index-embeddings-huggingface==0.5.4",
    "llama-index-readers-file>=0.1.0",
    "markdownify>=1.1.0",
    "modal>=1.0.4",
    "requests>=2.32.4",
    "tqdm>=4.66.0",
    "transformers==4.50.0",
    "torch==2.5.1",
    "unstructured>=0.10.0",
    "websocket-client>=1.8.0",
    "sphn>=0.2.0",
    "opuslib>=3.0.1",
    "websockets>=15.0.1",
    "html-to-markdown>=1.9.0",
    "lxml>=6.0.0",
    "beautifulsoup4>=4.13.4",
    "moviepy>=2.2.1",
    "matplotlib>=3.10.5",
    "pipecat-ai[webrtc,openai,silero,local-smart-turn,noisereduce,soundfile] >=0.0.91",
    "fastapi>=0.120.1",
]
